requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.1
matplotlib==3.8.3
seaborn==0.13.2
streamlit==1.32.0
streamlit-option-menu==0.3.12
streamlit-extras>=0.2.0
joblib>=1.1.0
plotly==5.19.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
xgboost==2.0.3
